// lib/guatecompras/types.ts
export interface paths {
// ...
    "/record/{ocid}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Obtiene una lista de RecordPackage */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    /** @description Es el identificador del record Default value : <b>ocds-xqjsxa-16024362</b> */
                    ocid: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/release/searchNog/{nog}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Obtine el último release, por identificador NOG (Número de operación de Guatecompras)
         *     Ejemplo: 16024362
         */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    /** @description NOG (Número de operación de Guatecompras) */
                    nog: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["HeaderRelease"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/release/search": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Búsqueda y lista de releases por ciertos criterios, recorrido secuencial (método Search alter) para explorar más de 10,000 resultados
         * @description **Tabla de parámetros:**
         *
         *     Parámetro | Descripción | Ejemplo
         *     --------- | ----------- | ----------
         *     anio | Año de la publicacion | 2021
         *     mes | Mes de la publicación | 2 dígitos => 10 (Octubre)
         *     dia | Día de la publicación | 2 dígitos => 25
         *     tipo_entidad | Tipo de entidad | dígitos
         *     sub_tipo_entidad | Sub tipo de entidad | dígitos
         *     entidad | Entidad compradora | dígitos
         *     unidad_compradora | Unidad compradora | dígitos
         *     modalidad_compradora | Modalidad de compra | dígitos
         *     sub_modalidad_compradora | Sub modalidad de compra | dígitos
         *     estatus_concurso | Estado de concurso | dígitos
         *
         *     **Nota:** parámetro de tiempo obligatorio (Año ó Mes ó Día)
         */
        get: {
            parameters: {
                query?: {
                    /** @description Obtiene un número de página específica del los Releases o Entregas */
                    pagina?: number;
                    /** @description Ordena los releases por la fecha publicación */
                    ordenamiento?: components["schemas"]["orderDatePublishNOG"];
                    /** @description Año de la publicacion */
                    Anio?: number;
                    /** @description Mes de la publicación */
                    Mes?: number;
                    /** @description Día de la publicación */
                    Dia?: number;
                    /** @description Tipo de entidad */
                    Tipo_entidad?: number;
                    /** @description Sub tipo de entidad */
                    Sub_tipo_entidad?: number;
                    /** @description Entidad compradora */
                    Entidad?: number;
                    /** @description Unidad compradora */
                    Unidad_compradora?: number;
                    /** @description Modalidad de compra */
                    Modalidad_compradora?: number;
                    /** @description Sub modalidad de compra */
                    Sub_modalidad_compradora?: number;
                    /** @description Estado de concurso */
                    Estatus_concurso?: number;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["HeaderRelease"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Awards: {
            id?: string | null;
            title?: string | null;
            status?: string | null;
            statusDetails?: string | null;
            date?: string | null;
            value?: components["schemas"]["Value"];
            suppliers?: components["schemas"]["Estructura_Id_Nombre"][] | null;
        };
        Bids: {
            details?: components["schemas"]["Details"][] | null;
        };
        ContactPoint: {
            name?: string | null;
            email?: string | null;
            telephone?: string | null;
            faxNumber?: string | null;
            url?: string | null;
        };
        Contracts: {
            id?: string | null;
            awardID?: string | null;
            title?: string | null;
            contractNumber?: string | null;
            status?: string | null;
            statusDetails?: string | null;
            dateSigned?: string | null;
            period?: components["schemas"]["TenderPeriod"];
            value?: components["schemas"]["Value"];
            documents?: components["schemas"]["Documents"][] | null;
            implementation?: components["schemas"]["Implementation"];
        };
// ...
        Parties: {
            name?: string | null;
            id?: string | null;
            identifier?: components["schemas"]["Identifier"];
            address?: components["schemas"]["Address"];
            contactPoint?: components["schemas"]["ContactPoint"];
            roles?: string[] | null;
            additionalContactPoints?: components["schemas"]["AdditionalContactPoints"][] | null;
            memberOf?: components["schemas"]["Estructura_Id_Nombre"][] | null;
            details?: components["schemas"]["DetailsTender"];
        };
        Releases: {
            id?: string | null;
            ocid?: string | null;
            date?: string | null;
            publishedDate?: string | null;
            language?: string | null;
            tag?: string[] | null;
            initiationType?: string | null;
            parties?: components["schemas"]["Parties"][] | null;
            buyer?: components["schemas"]["Estructura_Id_Nombre"];
            bids?: components["schemas"]["Bids"];
            tender?: components["schemas"]["Tender"];
            awards?: components["schemas"]["Awards"][] | null;
            contracts?: components["schemas"]["Contracts"][] | null;
            sources?: components["schemas"]["Sources"][] | null;
            dataSegmentation?: components["schemas"]["DataSegmentation"];
        };
// ...
        Tender: {
            id?: string | null;
            title?: string | null;
            status?: string | null;
            statusDetails?: string | null;
            datePublished?: string | null;
            procuringEntity?: components["schemas"]["Estructura_Id_Nombre"];
            items?: components["schemas"]["Items"][] | null;
            procurementMethod?: string | null;
            procurementMethodDetails?: string | null;
            mainProcurementCategory?: string | null;
            additionalProcurementCategories?: string[] | null;
            submissionMethod?: string[] | null;
            submissionMethodDetails?: string | null;
            tenderPeriod?: components["schemas"]["TenderPeriod"];
            /** Format: int32 */
            numberOfTenderers?: number;
            tenderers?: components["schemas"]["Estructura_Id_Nombre"][] | null;
            documents?: components["schemas"]["Documents"][] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}